feat: implement IsAllowType function and update related tests; remove duplicate implementation
