+++ internal/utils/utils.go
@@ -37,3 +37,26 @@ func ContainsFromArray(slice []string, item string) bool {
 func CheckPType(str string) bool {
 	return !strings.Contains(str, "toBlind") && !strings.Contains(str, "toGrepping")
 }
+
+// IsAllowType is checking content-type
+func IsAllowType(contentType string) bool {
+	notScanningType := map[string]struct{}{
+		"application/json":       {},
+		"application/javascript": {},
+		"text/javascript":        {},
+		"text/plain":             {},
+		"text/css":               {},
+		"image/jpeg":             {},
+		"image/png":              {},
+		"image/bmp":              {},
+		"image/gif":              {},
+		"application/rss+xml":    {},
+	}
+
+	for n := range notScanningType {
+		if strings.Contains(contentType, n) {
+			return false
+		}
+	}
+	return true
+}
